Inference

Basic algorithms

MarkovModels.pdfposteriors — Function

pdfposteriors(cfsm, lhs)
pdfposteriors(union(cfsm1, cfsm2, ...), batch_lhs)

Calculate the conditional posterior of "assigning" the $n$th frame to the $i$th pdf. The output is a tuple γ, ttl where γ is a matrix (# pdf x # frames) and ttl is the total probability of the sequence. This function can also be called in "batch-mode" by providing a union of compiled fsm and a 3D tensor containing the per-state, per-frame and per-batch values.

MarkovModels.maxstateposteriors — Function

maxstateposteriors(cfsm, lhs)

Calculate the posterior of "assigning" the $n$th frame to the $i$th state conditioned of all other states maximizing the likelihood of the sequence. The output is a matrix μ (# pdf x # frames).

MarkovModels.bestpath — Function

bestpath(cfsm, μ)

Return the sequence of state with the highest value from μ, i.e. the ouput of maxstateposteriors.