Semirings

Algorithms implemented in this package are very generic as they can be applied on various type of "numbers":

probabilities (values between $0$ and $+\infty$)

log-probabilities (values between $-\infty$ and $+\infty$)

...

Formally, they can operate on any semiring or semifield for some of them.

To work with these algebraic structures, the package defines the following abstract types:

MarkovModels.Semirings.Semiring — Type

abstract type Semiring <: Number end

Abstract base type for all semirings.

MarkovModels.Semirings.Semifield — Type

abstract type Semifield <: Semiring end

Abstract base type for all semifields.

Concrete types

The package provides the following concrete types:

MarkovModels.Semirings.LogSemifield — Type

struct LogSemifield{T<:AbstractFloat} <: Semfield
    val::T
end

Log-semifield is defined as :

$x \oplus y \triangleq \ln( e^x + e^y)$

$x \otimes y \triangleq x + y$

$x \oslash y \triangleq x - y$

$\forall x, y \in \mathbb{R}$.

MarkovModels.Semirings.ProbabilitySemifield — Type

struct ProbabilitySemifield{T<:AbstractFloat} <: Semfield
    val::T
end

Log-semifield is defined as :

$x \oplus y \triangleq x + y$

$x \otimes y \triangleq x \cdot y$

$x \oslash y \triangleq \frac{x}{y}$

$\forall x, y \in [0, 1]$.

MarkovModels.Semirings.TropicalSemiring — Type

struct TropicalSemiring{T<:AbstractFloat} <: Semiring
    val::T
end

Log-semifield is defined as :

$x \oplus y \triangleq \text{max}(x, y)$

$x \otimes y \triangleq x + y$

$\forall x, y \in \mathbb{R}$.