Inference

Baum-Welch algoritm (forward-backward)

The Baum-Welch algorithm computes the probability to be in a state i at time $n$:

\[p(z_n = i | x_1, ..., x_N)\]

It is implemented by the αβrecursion function.

MarkovModels.αβrecursion — Function

αβrecursion(cfsm, lhs::AbstractMatrix{T}; pruning::T = typemax(T) where T

Return the states log conditional probabilities (as a matrix) and the total likelihood of the sequence. pruning is a LogSemifield number between zero(T) and typemax(T). The lower it is the more pruning will occurs. Note that using pruning only makes sense if cfsm was compiled with allocator = spzeros.

See also: compile